Detect schemes and configurations more accurately
**The problem**
`xcodebuild` takes either a project or the combination of workspace + scheme as parameters.

When a project is passed, things are simple.

In case of workspace and scheme, there are a couple of issues:
* `xcodebuild -list` returns all the schemes, no matter local or shared, even Pods schemes. This leads to the situation when `Select Scheme:` lists a bunch of Pod schemes.
* `configurations` method returns an empty array because configurations are defined by projects, not by workspaces.

**Proposed solution**
Using the build configurations obtained by passing the workspace and scheme to `xcodebuild -showBuildSettings`, we find the project and  pass the project to `xcodebuild -list`.
Thus, `xcodebuild -list` returns just the relevant schemes and the configurations for the project.